split gamefield
gamefieldのファイルを分割した。
