+++ puyo/Assets/Editor/test_gamefiled.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections;
+using game_field;
+using NUnit.Framework;
+using UnityEditor;
+using UnityEngine;
+using UnityEngine.TestTools;
+
+public class test_gamefiled {
+
+	[Test]
+	public void test_init () {
+		//		GameField test_target = new GameField ();
+		//		Assert.AreEqual (0, 0);
+	}
+
+}
