+++ puyo/Assets/script/GameField_private.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 73abe5dcd1e643147ace21588a6b2b36
+timeCreated: 1515404259
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
